Rename test file to better reflect what to test
The current test is really only testing the casting of a std::any type
to a double in the update() method of the Observer. The tests should be
aimed at verifying the interface between the Publisher and the Observer
objects base classes - hence the change of file name to
PublisherObserverTests.cc. To test this interface will require both a
Publisher and an observer and some test doubles to implement
functionality from the base classes.

YouTrack Issue: ^PCalc-3