Set default value for error condition
If the default value is not set, it could be any value. This means the
test with a mock error using the default constructor does not set the
errorCondition explicitly (it only contains whatever happens to be in
that memory location at the time). Therefore, the test may pass or it
may fail during different test runs.

YouTrack Issue: ^PCalc-3